Update bulk-sg-run.py
I modified the summary to output passing and failing fonts by language tag to a JSON file. I also moved the missing tag data into this summary.

I was having trouble running a large number of fonts without the process being killed due to a shortage of memory. This may not be an issue on a more robust system.

I modified the script to slice up the available tags and hand those off to the Pool for processing in batches. This helps return some resources back to the system when the batches complete and processes shut down.